% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binomialCoefficient}
\alias{binomialCoefficient}
\title{Binomial Coefficient}
\usage{
binomialCoefficient(n, x)
}
\arguments{
\item{n}{Scalar of sample size}

\item{x}{Scalar of success count}
}
\value{
Coefficient for binomial pmf
}
\description{
Calculates the binomial coefficient
}
