% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binomialPMF}
\alias{binomialPMF}
\title{Binomial Probability Density}
\usage{
binomialPMF(n, x, p)
}
\arguments{
\item{n}{Scalar of sample size}

\item{x}{Scalar of success}

\item{p}{Scalar probability of success}
}
\value{
Probability density
}
\description{
Alternative to using dbinom
}
