% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackageTemplate.R
\docType{package}
\name{gmhs2}
\alias{gmhs2-package}
\alias{gmhs2}
\title{A template for Rcpp(Armadillo) packages}
\description{
An R package template illustrating the use of various packages including:
\code{\link[Rcpp]{Rcpp}}, \code{\link[RcppArmadillo]{RcppArmadillo}},
\code{\link[devtools]{devtools}}, \code{\link[roxygen2]{roxygen2}},
and \code{\link[testthat]{testthat}}.
}
