% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginalLikelihood}
\alias{marginalLikelihood}
\title{Calculates the marginal likelihood}
\usage{
marginalLikelihood(Y, mu, tau2, sim, a_gamma, b_gamma)
}
\arguments{
\item{Y}{A numeric matrix with columns for ric, nic, rit, nit}

\item{mu}{The mean}

\item{tau2}{The variance}

\item{sim}{The number of simulations}

\item{a_gamma}{The mean of prior distribution for gammas}

\item{b_gamma}{The variance of prior distribution for gammas}
}
\value{
Vector of likelihoods
}
\description{
Avoiding numeric integration of thetas and gammas
}
