% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.R
\name{mh}
\alias{mh}
\title{Metropolis-Hastings Single Iteration}
\usage{
mh(Y, mu, tau2, gammas_mat, thetas_mat, k, mh_scale, crit, hypers)
}
\arguments{
\item{Y}{Data table with columns rit, nit, ric, nic, theta_i, gamma_i}

\item{mu}{Estimate for mean}

\item{tau2}{Estimate for tau2}

\item{gammas_mat}{2xk matrix of current and pregenerated candidate gammas}

\item{thetas_mat}{2xk matrix of current and pregenerated candidate thetas}

\item{k}{The number of trials in Y}

\item{mh_scale}{MH algorithm variance}

\item{crit}{Critical value for updating}

\item{hypers}{Hyper parameters wity priors a_mu, b_mu, a_gamma, b_gamma, a_tau2, b_tau2}
}
\value{
list of new thetas and gammas
}
\description{
Metropolis-Hastings Single Iteration
}
\examples{
Y <- read.table(text = "
trial rit nit ric nic
Balcon 14 56 15 58
Clausen 18 66 19 64
Multicentre 15 100 12 95
Barber 10 52 12 47
Norris 21 226 24 228
Kahler 3 38 6 31
Ledwich 2 20 3 20
", header = TRUE)
Y$theta_i <- log(Y$rit/Y$nit/(1-Y$rit/Y$nit))-
log(Y$ric/Y$nic/(1-Y$ric/Y$nic))
Y$gamma_i<- 0.5*(log(Y$rit/Y$nit/(1-Y$rit/Y$nit))+
log(Y$ric/Y$nic/(1-Y$ric/Y$nic)))
hypers <- list(#Mu priors N(0,100)
"a_mu" = 0,
"b_mu" = 100,
#Gamma priors N(0,100)
"a_gamma" = 0,
"b_gamma" = 100,
#Tau2 priors
"a_tau2" = 0.001,
"b_tau2" = 0.001)

gammas_mat <- matrix(c(-1.0758811, -0.9215264, -1.8342675, -1.2527630,
-2.2092769, -1.9419261, -1.9659128,
-0.1591379, -1.3580273, -1.4989406,
0.0635526,  0.5562330, -0.5041676, -0.8886074), nrow = 2, ncol = 7,
byrow = TRUE)
thetas_mat <- matrix(c(-0.04546237, -0.11860574,  0.19933290, -0.36464311,
-0.13842138, -1.02961942, -0.46262352,
-1.9667471, -0.5626374,  0.3088954, -0.7944453,  1.3856203,  0.3028983,
 0.3382107), nrow = 2, ncol = 7,
byrow = TRUE)
mh(Y, -0.2, 0.07, gammas_mat, thetas_mat, 7, 0.23, -0.3925829, hypers)
}
