% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhGibbs.R
\name{mhGibbs}
\alias{mhGibbs}
\title{Metropolis-Hastings Gibbs Hybrid Algorithm}
\usage{
mhGibbs(
  Y,
  hypers,
  mh_scale,
  sim = 100,
  burn = 10,
  Bayesian = "hierarchical",
  hat_mu = hypers$a_mu,
  hat_tau2 = hypers$a_tau2
)
}
\arguments{
\item{Y}{Data matrix}

\item{hypers}{Vector of hyperparameters}

\item{mh_scale}{Scalar MH variance scale}

\item{sim}{Scalar number of simulations}

\item{burn}{Scalar number of burn in}

\item{Bayesian}{String either "hierarchical" or "empirical"}

\item{hat_mu}{Max likelihood for mu}

\item{hat_tau2}{Max likelihood for tau2}
}
\value{
Posteriors
}
\description{
Metropolis-Hastings Gibbs Hybrid Algorithm
}
\examples{
 Y <- read.table(text = "
trial rit nit ric nic
Balcon 14 56 15 58
Clausen 18 66 19 64
Multicentre 15 100 12 95
Barber 10 52 12 47
Norris 21 226 24 228
Kahler 3 38 6 31
Ledwich 2 20 3 20
", header = TRUE)
Y$theta_i <- log(Y$rit/Y$nit/(1-Y$rit/Y$nit))-
log(Y$ric/Y$nic/(1-Y$ric/Y$nic))
Y$gamma_i<- 0.5*(log(Y$rit/Y$nit/(1-Y$rit/Y$nit))+
log(Y$ric/Y$nic/(1-Y$ric/Y$nic)))
hypers <- list(
"a_mu" = 0,
"b_mu" = 100,
"a_gamma" = 0,
"b_gamma" = 100,
"a_tau2" = 0.001,
"b_tau2" = 0.001)

mhGibbs(Y, hypers, 0.23, sim = 100, burn = 10, "hierarchical")
mhGibbs(Y, hypers, 0.23, sim = 100, burn = 10, "empirical")
}
