% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postGamma.R
\name{postGamma}
\alias{postGamma}
\title{Posterior Gamma}
\usage{
postGamma(b_gamma, mu, tau2, theta, gamma, rt, rc)
}
\arguments{
\item{b_gamma}{Variance for the gamma param}

\item{mu}{Scalar mu estimate}

\item{tau2}{Scalar tau2 estimate}

\item{theta}{Scalar theta estimate}

\item{gamma}{Scalar gamma estimate}

\item{rt}{Scalar successes in the trial group}

\item{rc}{Scalar successes in the control group}
}
\value{
Updated Gamma
}
\description{
Posterior Gamma
}
\examples{
postGamma(100, 0.2, 0.15, theta = -0.05, gamma = -1.07, rt = 14, rc = 15)
}
