% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postMu.R
\name{postMu}
\alias{postMu}
\title{Posterior Mu}
\usage{
postMu(x, k, sigma, thetas, b_mu)
}
\arguments{
\item{x}{A value from a standard normal distribution to be transformed}

\item{k}{Scalar number of individual trials}

\item{sigma}{Scalar variance}

\item{thetas}{Vector of current estimates of length k}

\item{b_mu}{Scalar hyperparameter for the variance of mu}
}
\value{
A transformed single draw from a standard normal distribution with updated mean and variance
}
\description{
Posterior Mu
}
\examples{
postMu(x = 0.7,
        k = 7,
        sigma = 0.19,
        thetas = c(-0.05, -0.12, 0.20, -0.36, -0.14, -1.03, -0.46),
        b_mu = 100 )

}
