% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTau.R
\name{postTau}
\alias{postTau}
\title{Posterior Tau2}
\usage{
postTau(x, k, a_tau, b_tau, mu, thetas)
}
\arguments{
\item{x}{A value from a gamma distribution to be transformed}

\item{k}{Number of trials}

\item{a_tau}{Shape}

\item{b_tau}{Scalar hyperparameter for gamma distribution rate}

\item{mu}{Scalar mean}

\item{thetas}{Vector of current estimates of length k}
}
\value{
Inverse of a transformed draw from a gamma
}
\description{
Posterior Tau2
}
\examples{
postTau(x = 0.15,
        a_tau = 0.001,
        k = 7,
        b_tau = 0.001,
        mu = -0.2,
        thetas = c(-0.05, -0.12, 0.20, -0.36, -0.14, -1.03, -0.46))

}
