% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTheta.R
\name{postTheta}
\alias{postTheta}
\title{Posterior Theta}
\usage{
postTheta(tau2, mu, gamma, theta, rt, rc)
}
\arguments{
\item{tau2}{Current tau2 estimate}

\item{mu}{Current mu estimate}

\item{gamma}{Scalar gamma}

\item{theta}{Scalar gamma estimate}

\item{rt}{Scalar successes in the trial group}

\item{rc}{Scalar successes in the control group}
}
\value{
Updated Gamma
}
\description{
Posterior Theta
}
\examples{
postTheta(0.07, -0.2, gamma = -1.34, theta = -0.05, rt = 14, rc = 15)


}
